+++ docs/web/running_the_app.md
@@ -1,0 +1,5 @@
+# Locally
+Need to have python
+
+
+# Deployment
